Time: 0 ms (100%), Space: 8.4 MB (82.61%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        int rev= 0;
+        long long rev= 0;
         while(x!=0){
             int lastdigit = x % 10;
              rev = rev * 10 + lastdigit;
